+++ client/package.json
@@ -12,6 +12,7 @@
     "dependencies": {
         "@geist-ui/core": "^2.3.8",
         "next": "12.3.0",
+        "node-fetch": "^3.2.10",
         "react": "18.2.0",
         "react-dom": "18.2.0"
     },
@@ -19,6 +20,7 @@
         "@types/node": "18.7.17",
         "@types/react": "18.0.19",
         "@types/react-dom": "18.0.6",
+        "dotenv": "^16.0.3",
         "eslint": "8.23.1",
         "eslint-config-next": "12.3.0",
         "typescript": "4.8.3"
