+++ client/pages/[...view-skill].tsx
