+++ server/package.json
@@ -13,6 +13,7 @@
     "dependencies": {
         "body-parser": "^1.20.0",
         "celebrate": "^15.0.1",
+        "cors": "^2.8.5",
         "dotenv": "^16.0.2",
         "express": "^4.18.1",
         "mysql2": "^2.3.3",
@@ -22,6 +23,7 @@
         "umzug": "^3.2.1"
     },
     "devDependencies": {
+        "@types/cors": "^2.8.12",
         "@types/express": "^4.17.13",
         "@types/jest": "^29.0.1",
         "@types/node": "^18.7.17",
