+++ server/src/lib/models/Role.ts
@@ -1,19 +1,28 @@
-import { DataTypes } from 'sequelize';
+import { DataTypes, ModelDefined, Optional } from 'sequelize';
 import { sequelize } from '../../database';
 
-export const Role = sequelize.define('role', {
-    Role_ID: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-    },
-    Role_Name: {
-        type: DataTypes.STRING,
-        unique: true,
-        allowNull: false,
-    },
-    Role_Description: {
-        type: DataTypes.STRING,
-        allowNull: true,
-    },
-});
+export interface RoleAttributes {
+    Role_ID: string;
+    Role_Name: string;
+    Role_Description: string;
+}
+
+type RoleCreationAttributes = Optional<RoleAttributes, 'Role_ID'>;
+
+export const Role: ModelDefined<RoleAttributes, RoleCreationAttributes> =
+    sequelize.define('role', {
+        Role_ID: {
+            type: DataTypes.INTEGER,
+            autoIncrement: true,
+            primaryKey: true,
+        },
+        Role_Name: {
+            type: DataTypes.STRING,
+            unique: true,
+            allowNull: false,
+        },
+        Role_Description: {
+            type: DataTypes.STRING,
+            allowNull: true,
+        },
+    });
