+++ server/src/routes/roles.ts
@@ -16,10 +16,7 @@ roles.post(
     async (req, res) => {
         try {
             const Role_Name = req.body.Role_Name.trim();
-            let Role_Description;
-            if (req.body.Role_Description) {
-                Role_Description = req.body.Role_Description.trim();
-            }
+            const Role_Description = (req.body.Role_Description ?? '').trim();
 
             if (Role_Name.length === 0) {
                 throw new Error('Name of roles must be specified.');
@@ -58,14 +55,22 @@ roles.put(
         },
         body: {
             Role_Name: Joi.string().required(),
+            Role_Description: Joi.string(),
         },
     }),
 
     async (req, res) => {
         try {
-            const Role_Name = req.body.Role_Name.trim();
             const { Role_ID } = req.params;
             const role = await Role.findByPk(Role_ID);
+
+            const Role_Name =
+                req.body.Role_Name.trim() ?? role?.get('Role_Name') ?? '';
+            const Role_Description = (
+                req.body.Role_Description ??
+                '' ??
+                role?.get('Role_Description')
+            ).trim();
 
             if (!role) {
                 return res.status(404).json({ error: 'Role not found' });
@@ -75,7 +80,10 @@ roles.put(
                 throw new Error('Name of roles must be specified.');
             }
 
-            await Role.update({ Role_Name: Role_Name }, { where: { Role_ID } });
+            await Role.update(
+                { Role_Name, Role_Description },
+                { where: { Role_ID } }
+            );
 
             res.json({ Role_ID, Role_Name });
         } catch (error) {
