+++ server/tsconfig.json
@@ -16,6 +16,7 @@
         "strictPropertyInitialization": true,
         "outDir": "dist",
         "baseUrl": ".",
+        "esModuleInterop": true,
         "paths": {
             "@routes/*": ["./src/routes/*"],
             "@lib/*": ["./src/lib/*"]
